backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
python-multipart==0.0.6
joblib==1.3.2
pillow==10.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
redis==5.0.1
celery==5.3.4
httpx==0.25.2
structlog==23.2.0
sentry-sdk==1.38.0
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1

# GraphQL
strawberry-graphql==0.235.0
graphql-core==3.2.3

# WebSocket
websockets==12.0

# IPFS
ipfshttpclient==0.8.0a2

# TimescaleDB
timescaledb==0.1.0

# File storage
boto3==1.34.0

# Message queue
pika==1.3.2

# Monitoring
prometheus-client==0.19.0

# Background tasks
apscheduler==3.10.4

# Email templates
jinja2==3.1.2

# Image processing
opencv-python==4.8.1.78

# Geospatial
geopy==2.4.0
shapely==2.0.2

# Blockchain
web3==6.15.0
eth-account==0.10.0

# Additional blockchain libraries
eth-abi==4.2.1
eth-utils==2.3.1
hexbytes==0.3.1